More strict code checking

+++ src/sm/sm-eac.c
@@ -250,7 +250,7 @@ static int encode_mse_cdata(struct sc_context *ctx, int protocol,
 	memcpy(p, data, data_len);
 	memcpy(p+data_len, encoded_chat, encoded_chat_len);
 	*cdata = p;
-	r = data_len + encoded_chat_len;
+	r = (int)(data_len + encoded_chat_len);
 
 	sc_debug_hex(ctx, SC_LOG_DEBUG_SM, "MSE command data", p, r);
 
@@ -405,7 +405,7 @@ static int encode_gen_auth_cdata(struct sc_context *ctx,
 
 	r = sc_asn1_encode(ctx, capdu_eac_gen_auth, cdata, &data_len);
 	SC_TEST_GOTO_ERR(ctx, SC_LOG_DEBUG_VERBOSE, r, "Error encoding General Authenticate APDU data");
-	r = data_len;
+	r = (int)data_len;
 
 err:
 	return r;
@@ -1553,7 +1553,7 @@ eac_sm_encrypt(sc_card_t *card, const struct iso_sm_ctx *ctx,
 	}
 	*enc = p;
 	memcpy(*enc, encbuf->data, encbuf->length);
-	r = encbuf->length;
+	r = (int)encbuf->length;
 
 err:
 	BUF_MEM_clear_free(databuf);
@@ -1594,7 +1594,7 @@ eac_sm_decrypt(sc_card_t *card, const struct iso_sm_ctx *ctx,
 	}
 	*data = p;
 	memcpy(*data, databuf->data, databuf->length);
-	r = databuf->length;
+	r = (int)databuf->length;
 
 err:
 	BUF_MEM_clear_free(databuf);
@@ -1641,7 +1641,7 @@ eac_sm_authenticate(sc_card_t *card, const struct iso_sm_ctx *ctx,
 	}
 	*macdata = p;
 	memcpy(*macdata, macbuf->data, macbuf->length);
-	r = macbuf->length;
+	r = (int)macbuf->length;
 
 err:
 	if (inbuf)
